code opgeschoond

+++ src/main/java/com/example/eindopdrachtyarnicornback/Service/ProfileService.java
@@ -4,15 +4,12 @@ import com.example.eindopdrachtyarnicornback.DTO.ProfileAndUserDto;
 import com.example.eindopdrachtyarnicornback.DTO.ProfileDto;
 import com.example.eindopdrachtyarnicornback.Exceptions.IdNotFoundException;
 import com.example.eindopdrachtyarnicornback.Models.Profile;
-import com.example.eindopdrachtyarnicornback.Models.User;
 import com.example.eindopdrachtyarnicornback.Repository.ProfileRepository;
 import com.example.eindopdrachtyarnicornback.Repository.RoleRepository;
 import com.example.eindopdrachtyarnicornback.Repository.UserRepository;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.RequestBody;
-
-import com.example.eindopdrachtyarnicornback.Models.Role;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -21,14 +18,10 @@ import java.util.Optional;
 public class ProfileService {
 
     private final ProfileRepository profileRepository;
-    private final PasswordEncoder passwordEncoder;
-    private final UserRepository userRepository;
-    private final RoleRepository roleRepository;
+
+
     public ProfileService(ProfileRepository profileRepository, PasswordEncoder passwordEncoder, UserRepository userRepository, RoleRepository roleRepository) {
         this.profileRepository = profileRepository;
-        this.passwordEncoder = passwordEncoder;
-        this.userRepository = userRepository;
-        this.roleRepository = roleRepository;
     }
 
 
@@ -84,15 +77,20 @@ public class ProfileService {
         return savedProfileDto;
     }
 
-    public String deleteProfile(@RequestBody Long id) {
-        profileRepository.deleteById(id);
-
-        return "Profile deleted";
-    }
     private void profileAndUserDtoToProfile(ProfileAndUserDto pDto, Profile p) {
         p.setFirstName(pDto.getFirstName());
         p.setLastName(pDto.getLastName());
         p.setEmail(pDto.getEmail());
+    }
+
+    public String deleteProfile(@RequestBody Long id) {
+        if (profileRepository.existsById(id)) {
+            profileRepository.deleteById(id);
+        } else {
+            throw new IdNotFoundException("Profile not found with ID: " + id);
+        }
+
+        return "Profile deleted";
     }
 
 }
